+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.1.0] / 2022-12-06
 ### Features
 - Feature `TestEngine.GetTestFullName` to compare with `TestEngine.GetTestFullNames`.
+### Fixed
+- Fix Multiple AssertionResults: `TestExecutionContext.CurrentContext.CurrentResult.AssertionResults.Clear();`
 
 ## [1.0.9] / 2022-12-06
 ### Features